fixing mask projection

+++ util/utilIO.cpp
@@ -45,7 +45,7 @@ void MeshIO::saveChngMask3d(const std::vector<cv::Mat> &pts_3d, const std::strin
   MyMesh m;
 
   cv::Mat tmpMat;
-  double w, x, y, z;
+  float w, x, y, z;
 
   for(int i = 2 ; i < 3 ; i++){
     DrawProgressBar(40, (double)i/(double)pts_3d.size());
@@ -54,10 +54,10 @@ void MeshIO::saveChngMask3d(const std::vector<cv::Mat> &pts_3d, const std::strin
       
       tmpMat  = pts_3d[i].col(c);
 
-      w = tmpMat.at<double>(3,0);
-      x = tmpMat.at<double>(0,0)/w;
-      y = tmpMat.at<double>(1,0)/w;
-      z = tmpMat.at<double>(2,0)/w;
+      w = tmpMat.at<float>(3,0);
+      x = tmpMat.at<float>(0,0)/w;
+      y = tmpMat.at<float>(1,0)/w;
+      z = tmpMat.at<float>(2,0)/w;
 
       vcg::tri::Allocator<MyMesh>::AddVertex(m, MyMesh::CoordType(x,y,z));
